Add token to track/untrack/update links.

+++ user/untrack.php
@@ -10,6 +10,11 @@ if (!$user->valid()) {
   exit;
 }
 
+if ($_REQUEST['token'] != $user->token()) {
+  echo "Invalid token\n";
+  exit;
+}
+
 $sql = "delete from f_tracking where fid = " . $forum['fid'] . " and tid = '" . addslashes($tid) . "' and aid = '" . $user->aid . "'";
 mysql_query($sql) or sql_error($sql);
